feat: sort channels returned from HAA
