+++ apps/microsoft-teams/app-actions/src/helpers/get-channels-list.ts
@@ -1,5 +1,7 @@
 import { config } from '../config';
-import { AppActionCallResponse, Channel, TeamInstallation } from '../types';
+import { Channel, TeamInstallation } from '../types';
+
+const GENERAL_CHANNEL_NAME = 'general';
 
 export const getChannelsList = async (tenantId: string): Promise<Channel[]> => {
   const response = await config.msTeamsBotService.getTeamInstallations(tenantId);
@@ -9,7 +11,8 @@ export const getChannelsList = async (tenantId: string): Promise<Channel[]> => {
   }
 
   const channelsList = transformInstallationsToChannelsList(response.data, tenantId);
-  return channelsList;
+  const sortedChannelsList = channelsList.sort(sortChannels);
+  return sortedChannelsList;
 };
 
 export const transformInstallationsToChannelsList = (
@@ -39,4 +42,26 @@ export const transformInstallationsToChannelsList = (
   }, [] as Channel[]);
 
   return channelsList;
+};
+
+/**
+ * Sorts the channels alpha by team name, then General channel, then alpha by channel name
+ */
+export const sortChannels = (a: Channel, b: Channel) => {
+  const teamA = a.teamName.toLowerCase();
+  const teamB = b.teamName.toLowerCase();
+  const channelA = a.name.toLowerCase();
+  const channelB = b.name.toLowerCase();
+
+  if (teamA === teamB) {
+    if (channelA === GENERAL_CHANNEL_NAME) {
+      return -1;
+    } else if (channelB === GENERAL_CHANNEL_NAME) {
+      return 1;
+    } else {
+      return channelA.localeCompare(channelB);
+    }
+  } else {
+    return teamA.localeCompare(teamB);
+  }
 };
