--- apps/microsoft-teams/app-actions/test/fixtures/mockChannels.ts
@@ -1,32 +1,0 @@
-import { Channel } from '../../src/types';
-
-export const mockChannels: Channel[] = [
-  {
-    id: '19:e3a386bd1e0f4e00a286b4e86b0cfbe9@thread.tacv2',
-    name: 'General',
-    teamId: 'ed57f808-c14f-4a53-bf53-e36de0783385',
-    teamName: 'Marketing Team',
-    tenantId: '666e56a6-1f2a-47c7-b88c-1ed9e1bb8668',
-  },
-  {
-    id: '19:e2a385bd1e0f4e00a286b4e86b0cfbe9@thread.tacv2',
-    name: 'Branding',
-    teamId: 'ed57f808-c14f-4a53-bf53-e36de0783385',
-    teamName: 'Marketing Team',
-    tenantId: '666e56a6-1f2a-47c7-b88c-1ed9e1bb8668',
-  },
-  {
-    id: '19:39ca79ab85df4520af8a459bd1abaea1@thread.tacv2',
-    name: 'Corporate Marketing',
-    teamId: 'ed57f808-c14f-4a53-bf53-e36de0783385',
-    teamName: 'Marketing Team',
-    tenantId: '666e56a6-1f2a-47c7-b88c-1ed9e1bb8668',
-  },
-  {
-    id: '19:3bccfda604454e63bd839399e6752ba3@thread.tacv2',
-    name: 'General',
-    teamId: '1a91e6ef-ac80-4b9b-9989-d3416c38671c',
-    teamName: 'Sales Team',
-    tenantId: '666e56a6-1f2a-47c7-b88c-1ed9e1bb8668',
-  },
-];
